+++ app/components/design_system/nhsuk/summary_card_component.rb
@@ -1,0 +1,6 @@
+module DesignSystem
+  module Nhsuk
+    class SummaryCardComponent < DesignSystem::Generic::SummaryCardComponent
+    end
+  end
+end